+++ src/utils/redux.ts
@@ -3,7 +3,7 @@ import { createAsyncThunk as baseCreateAsyncThunk } from '@reduxjs/toolkit';
 import { identity } from 'ramda';
 import type { RootState } from '../container/store';
 
-export const createAsyncThunk = <Returned, ThunkArg>(
+export const createAsyncThunk = <Returned, ThunkArg = void>(
   typePrefix: string,
   payloadCreator: AsyncThunkPayloadCreator<Returned, ThunkArg, { state: RootState, serializedErrorType: any }>,
 ) => baseCreateAsyncThunk(