+++ src/visits/reducers/orphanVisits.ts
@@ -2,6 +2,7 @@ import type { ShlinkApiClient, ShlinkOrphanVisit, ShlinkOrphanVisitType } from '
 import { isBetween } from '../../utils/dates/helpers/date';
 import { isOrphanVisit } from '../types/helpers';
 import { createVisitsAsyncThunk, createVisitsReducer, lastVisitLoaderForLoader } from './common';
+import type { deleteOrphanVisits } from './orphanVisitsDeletion';
 import type { LoadVisits, VisitsInfo } from './types';
 
 const REDUCER_PREFIX = 'shlink/orphanVisits';
@@ -42,10 +43,14 @@ export const getOrphanVisits = (apiClientFactory: () => ShlinkApiClient) => crea
 
 export const orphanVisitsReducerCreator = (
   asyncThunkCreator: ReturnType<typeof getOrphanVisits>,
+  deleteOrphanVisitsThunk: ReturnType<typeof deleteOrphanVisits>,
 ) => createVisitsReducer({
   name: REDUCER_PREFIX,
   initialState,
   asyncThunkCreator,
+  extraReducers: (builder) => {
+    builder.addCase(deleteOrphanVisitsThunk.fulfilled, (state) => ({ ...state, visits: [] }));
+  },
   filterCreatedVisits: ({ query = {} }, createdVisits) => {
     const { startDate, endDate } = query;
     return createdVisits.filter(({ visit, shortUrl }) => !shortUrl && isBetween(visit.date, startDate, endDate));